PPTP-1423 Update partnership - Add partners (Post registration)
