+++ conf/application.conf
@@ -180,6 +180,7 @@ urls {
   groupMemberGrsCallback = "http://localhost:8503/register-for-plastic-packaging-tax/group-member-grs-callback"
   amendGroupMemberGrsCallback = "http://localhost:8503/register-for-plastic-packaging-tax/amend-add-group-member-grs-callback"
   partnerGrsCallback = "http://localhost:8503/register-for-plastic-packaging-tax/partner-grs-callback"
+  amendPartnerGrsCallback = "http://localhost:8503/register-for-plastic-packaging-tax/amend-add-partner-grs-callback"
   exitSurvey = "http://localhost:9514/feedback/plastic-packaging-tax-registration"
   feedback = {
     authenticatedLink = "http://localhost:9250/contact/beta-feedback"
